perf: cache passkey auth storage fixture envelope
